requirements.txt
streamlit==1.32.0
streamlit-authenticator==0.2.3
pandas==2.2.1
plotly==5.18.0
openpyxl==3.1.2
pyyaml==6.0.1
python-dotenv==1.0.1
loguru==0.7.2
bcrypt==4.0.1 